+++ src/Compute/Compute/ChangeLog.md
@@ -20,7 +20,7 @@
 
 -->
 ## Upcoming Release
-* New parameter `VM` in new parameter set `VMParameterSet` added to `Get-AzVMDSCExtensionStatus` cmdlet. 
+* New parameter `VM` in new parameter set `VMParameterSet` added to `Get-AzVMDscExtensionStatus` and `Get-AzVMDscExtension` cmdlets. 
 
 ## Version 4.7.0
 * Edited Get-AzVm to filter by `-Name` prior to checking for throttling due to too many resources.